pyproject.toml
[tool.poetry]
name = "whatsapp-apis"
version = "0.1.0"
description = ""
authors = ["Michkail <55208634+Michkail@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "whatsapp_apis"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "4.1"
djangorestframework = "^3.14.0"
markdown = "^3.5.1"
django-filter = "^23.4"
selenium = "^4.15.2"
trio = "0.22.0"
python-decouple = "^3.8"
channels = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
